+++ es/api/load.js
@@ -3,7 +3,7 @@ import util from '../support/util';
 
 /**
  * load will call startup
- * @param {{ [moduleName:string]: config:{state:object, reducer:object, watch:object, computed:object, init:object, sharedToGlobalMapping:object, isClassSingle:boolean} }} store
+ * @param {{ [moduleName:string]: config:{state:object, reducer:object, watch:object, computed:object, init:object, isClassSingle:boolean} }} store
  * @param option
  */
 
@@ -20,18 +20,16 @@ export default function (store = {}, option = {}) {
   const _watch = {};
   const _computed = {};
   let _init = {};
-  const _sharedToGlobalMapping = {};
   const _moduleSingleClass = {};
   const moduleNames = Object.keys(store);
   moduleNames.forEach(m => {
     const config = store[m];
-    const { state, reducer, watch, computed, init, sharedToGlobalMapping, isClassSingle } = config;
+    const { state, reducer, watch, computed, init, isClassSingle } = config;
     if (state) _store[m] = state;
     if (reducer) _reducer[m] = reducer;
     if (watch) _watch[m] = watch;
     if (computed) _computed[m] = computed;
     if (init) _init[m] = init;
-    if (sharedToGlobalMapping) _sharedToGlobalMapping[m] = sharedToGlobalMapping;
     if (typeof isClassSingle === 'boolean') _moduleSingleClass[m] = isClassSingle;
   });
 
@@ -42,7 +40,6 @@ export default function (store = {}, option = {}) {
     watch: _watch,
     computed: _computed,
     init: _init,
-    sharedToGlobalMapping: _sharedToGlobalMapping,
     moduleSingleClass: _moduleSingleClass,
   }