+++ es/core/base/index.js
@@ -3,7 +3,6 @@ export { default as buildCcClassContext } from './build-cc-class-context';
 export { default as catchCcError } from './catch-cc-error';
 export { default as computeCcUniqueKey } from './compute-cc-unique-key';
 export { default as getFeatureStrAndStpMapping } from './get-feature-str-and-stpmapping';
-export { default as handleModuleSharedToGlobalMapping } from './handle-module-shared-to-global-mapping';
 export { default as runLater } from './run-later';
 export { default as uuid } from './uuid';
 export { default as buildMockEvent } from './build-mock-event';