+++ src/cc-context/index.js
@@ -208,7 +208,7 @@ const ccContext = {
   refs,
   info: {
     startupTime: Date.now(),
-    version: '1.238.1',
+    version: '1.3.0',
     author: 'fantasticsoul',
     emails: ['624313307@qq.com', 'zhongzhengkai@gmail.com'],
     tag: 'xenogear',